MODIFIED:Added exceptions

+++ src/main/java/com/ammu/services/NotesServiceImpl.java
@@ -1,8 +1,8 @@
 package com.ammu.services;
 
 import java.time.LocalDateTime;
-import java.util.ArrayList;
 import java.util.List;
+import java.util.Optional;
 
 import javax.transaction.Transactional;
 
@@ -13,15 +13,22 @@ import org.springframework.core.env.Environment;
 import org.springframework.stereotype.Service;
 
 import com.ammu.dto.NotesDto;
-import com.ammu.dto.RemainderDto;
+import com.ammu.exceptionHandling.NoteIdNotFoundException;
 import com.ammu.model.NotesModel;
+import com.ammu.model.UserModel;
 import com.ammu.repository.NoteRepository;
+import com.ammu.repository.UserRepository;
 import com.ammu.response.Response;
+import com.ammu.security.utility.JwtToken;
 
 @Service
 @PropertySource("classpath:message.properties")
 public class NotesServiceImpl implements NotesService 
 {
+
+	@Autowired
+	UserRepository userRepository;
+
 	@Autowired
 	NoteRepository noteRepo;
 
@@ -38,24 +45,32 @@ public class NotesServiceImpl implements NotesService
 	Environment enviroment;
 
 	@Override
-	public Response addNote(NotesDto notesDto) 
+	public Response addNote(NotesDto notesDto , String token) 
 	{
-		mapper.map(notesDto , notesModel);
-		noteRepo.save(notesModel);
+		int id = JwtToken.retrieveIdFromToken(token);
+		Optional<UserModel> userData = userRepository.findById(id);
+		NotesModel noteData = mapper.map(notesDto , NotesModel.class);
+		noteData.setUserModel(userData.get());
+		noteRepo.save(noteData);
 		return new Response(enviroment.getProperty("noteAdd.success.text") , enviroment.getProperty("noteAdd.success.code"));
 	}
 
 	@Override
-	public Response deleteNote(Long id)
+	public Response deleteNote(String token)
 	{
+		int id = JwtToken.retrieveIdFromToken(token);
+		if(id == 0)
+			throw new NoteIdNotFoundException("This ID is not present","400" );
+		//<UserModel> userData = userRepository.findById(id);
 		noteRepo.deleteById(id);
 		return new Response(enviroment.getProperty("noteDelete.success.text") , enviroment.getProperty("noteDelete.success.code"));
 	}
 
-
+	@Transactional
 	@Override
-	public Response updateNote(NotesDto notesDto , Long id)
+	public Response updateNote(NotesDto notesDto , int id , String token) 
 	{
+
 		noteRepo.updateTitleAndDescription(notesDto.getTitle(), notesDto.getDescription(), id);
 		mapper.map(notesDto, notesModel);
 		return new Response(enviroment.getProperty("noteUpdate.success.text") , enviroment.getProperty("noteUpdate.success.code"));	
@@ -64,11 +79,16 @@ public class NotesServiceImpl implements NotesService
 	@Override
 	public List<NotesModel> retriveAllNote() 
 	{
+		//		 User user = userData.get();
+		//	        List<Note> allNotes = noteRepository.findAllByUser(user).stream()
+		//	                .filter(u -> !(u.isArchived() || u.isTrashed())).collect(Collectors.toList());
+		//	        return allNotes;
+		
 		return (List<NotesModel>) noteRepo.findAll();
 	}
 
 	@Override
-	public Response findByTitle(String title)
+	public Response findByTitle(String title , String Token) 
 	{
 		notesModel = noteRepo.findByTitle(title);
 		if(title == notesModel.getTitle())
@@ -79,7 +99,7 @@ public class NotesServiceImpl implements NotesService
 	}
 
 	@Override
-	public Response findByDescription(String description) 
+	public Response findByDescription(String description)
 	{
 		notesModel = noteRepo.findByDescription(description);
 		if(description == notesModel.getDescription());
